CLI-120 Move command under a dedicated directory
